WordFileParser support UTF-8
