+++ CMakeLists.txt
@@ -12,6 +12,7 @@ qt_add_executable(appwords_statistics
     main.cpp
     wordsstatisticsmodel.cpp
     controller.cpp
+    fileparser.cpp
 )
 
 qt_add_qml_module(appwords_statistics
